feat(web): extract job listing helpers into jobs-lib
